+++ examples/echo_component_bot.php
@@ -54,7 +54,7 @@ $comp = new JAXL(array(
     'host' => $argv[3],
     'port' => $argv[4],
 
-    'log_level' => JAXL_INFO
+    'log_level' => JAXLLogger::INFO
 ));
 
 //