+++ examples/echo_unix_sock_server.php
@@ -42,7 +42,7 @@ if ($argc < 2) {
 }
 
 require_once 'jaxl.php';
-JAXLLogger::$level = JAXL_INFO;
+JAXLLogger::$level = JAXLLogger::INFO;
 
 $server = null;