+++ examples/subscriber.php
@@ -48,7 +48,7 @@ require_once 'jaxl.php';
 $client = new JAXL(array(
     'jid' => $argv[1],
     'pass' => $argv[2],
-    'log_level' => JAXL_INFO
+    'log_level' => JAXLLogger::INFO
 ));
 
 $client->require_xep(array(