+++ examples/xmpp_rest.php
@@ -48,7 +48,7 @@ require_once 'jaxl.php';
 $xmpp = new JAXL(array(
     'jid' => $argv[1],
     'pass' => $argv[2],
-    'log_level' => JAXL_INFO
+    'log_level' => JAXLLogger::INFO
 ));
 
 // register callbacks on required xmpp events